requirements.txt
amqp==5.1.1
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo;python_version<"3.9"
billiard==4.1.0
binaryornot==0.4.4
Brotli==1.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cookiecutter==2.4.0
cron-descriptor==1.4.0
cssselect2==0.7.0
Django==4.2.5
django-admin-interface==0.26.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-colorfield==0.10.1
django-taggit==4.0.0
django-tailwind==3.6.0
django-timezone-field==6.0.1
django-tinymce==3.6.1
fonttools==4.43.1
greenlet==3.0.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
Jinja2==3.1.2
kombu==5.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
packaging==23.2
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2==2.9.9
pycparser==2.21
pydyf==0.8.0
Pygments==2.16.1
pyphen==0.14.0
pypng==0.20220715.0
python-barcode==0.15.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
rich==13.6.0
six==1.16.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
text-unidecode==1.3
tinycss2==1.2.1
types-python-dateutil==2.8.19.14
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
weasyprint==60.1
webencodings==0.5.1
zopfli==0.2.3
